% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_dtm.R
\name{dtm_obter_exclusoes}
\alias{dtm_obter_exclusoes}
\title{Obter tabela de exclusões de uma DTM}
\usage{
dtm_obter_exclusoes(dtm)
}
\arguments{
\item{dtm}{DocumentTermMatrix gerada por `cria_dtm`.}
}
\value{
Tibble contendo `doc_id` e `motivo` da exclusão.
}
\description{
Obter tabela de exclusões de uma DTM
}
