% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_referencia_nacional.R
\name{is_referencia_nacional}
\alias{is_referencia_nacional}
\title{Verifica se o artigo é de origem nacional com base no DOI}
\usage{
is_referencia_nacional(doi)
}
\arguments{
\item{doi}{String com o DOI do artigo.}
}
\value{
TRUE se for nacional, FALSE caso contrário.
}
\description{
Utiliza o redirecionamento do DOI para verificar se o artigo está hospedado em um domínio nacional,
e também verifica o publisher via CrossRef como fallback.
}
