% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquema.R
\name{schema_escrever}
\alias{schema_escrever}
\title{Gravar instrumento JSON}
\usage{
schema_escrever(schema, caminho, pretty = TRUE)
}
\arguments{
\item{schema}{Instrumento válido.}

\item{caminho}{Caminho de saída.}

\item{pretty}{JSON indentado.}
}
\value{
Caminho, de forma invisível.
}
\description{
Gravar instrumento JSON
}
