% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquema.R
\name{schema_ler}
\alias{schema_ler}
\title{Ler instrumento de arquivo JSON}
\usage{
schema_ler(caminho, error = TRUE)
}
\arguments{
\item{caminho}{Caminho do arquivo.}

\item{error}{Interrompe em erro de leitura ou validação.}
}
\value{
Instrumento como lista.
}
\description{
Ler instrumento de arquivo JSON
}
