% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquema.R
\name{schema_padrao}
\alias{schema_padrao}
\title{Instrumento geral configurável para ciências sociais}
\usage{
schema_padrao()
}
\value{
Lista compatível com o formato JSON de instrumento 1.0.0.
}
\description{
Instrumento geral configurável para ciências sociais
}
