% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquema.R
\name{schema_para_json}
\alias{schema_para_json}
\title{Serializar um instrumento para JSON}
\usage{
schema_para_json(schema, pretty = TRUE)
}
\arguments{
\item{schema}{Instrumento válido.}

\item{pretty}{Indica se o JSON deve ser indentado.}
}
\value{
String JSON.
}
\description{
Serializar um instrumento para JSON
}
