% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valida_evidencias.R
\name{valida_evidencias_resposta}
\alias{valida_evidencias_resposta}
\title{Validar conjunto de evidências vinculadas a uma resposta}
\usage{
valida_evidencias_resposta(resposta, evidencias, corpus = NULL, error = FALSE)
}
\arguments{
\item{resposta}{Lista de resposta compatível com contrato F0.}

\item{evidencias}{Lista de objetos de evidência.}

\item{corpus}{Opcionalmente, lista de passagens.}

\item{error}{Se `TRUE`, interrompe se houver incoerência entre resposta e evidências.}
}
\value{
Lista com `valid` e `errors`.
}
\description{
Validar conjunto de evidências vinculadas a uma resposta
}
