% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_pdf_scihub.R
\name{baixa_pdf_scihub}
\alias{baixa_pdf_scihub}
\title{Baixa o PDF de um artigo acadêmico usando o Sci-Hub, a partir de um DOI ou título}
\usage{
baixa_pdf_scihub(referencia, diretorio = ".", delay = 2)
}
\arguments{
\item{referencia}{Uma lista com pelo menos os campos `title` (obrigatório) e `doi` (opcional). O campo `id` será usado como nome do arquivo salvo.}

\item{diretorio}{Diretório onde o PDF será salvo. Será criado caso não exista. Padrão: `"."`.}

\item{delay}{Tempo (em segundos) de espera após cada download, para evitar bloqueios. Padrão: `2`.}
}
\value{
Nenhum valor de retorno. O PDF será salvo no diretório especificado.
}
\description{
Esta função acessa o Sci-Hub com base no DOI ou título da referência, extrai o link do PDF e realiza o download para o diretório especificado.
}
\examples{
\dontrun{
ref_com_doi <- list(
id = "exemplo_doi",
title = "Quid Pro Quo: Builders, Politicians, and Election Finance in India",
doi = "10.2139/ssrn.1972987" )
ref_sem_doi <- list(
  id = "exemplo_sem_doi",
    title = "Quid pro quo: Builders, politicians, and election finance in India")

baixa_pdf_scihub(ref_com_doi, diretorio = "pdfs")
baixa_pdf_scihub(ref_sem_doi, diretorio = "pdfs") }
}
