% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref_info.R
\name{crossref_info}
\alias{crossref_info}
\title{Consulta metadados de um artigo via CrossRef}
\usage{
crossref_info(doi)
}
\arguments{
\item{doi}{String com o DOI do artigo.}
}
\value{
Uma lista com os campos \code{publisher}, \code{container}, \code{type} e \code{url}.
}
\description{
Busca metadados como \code{publisher}, \code{type} e \code{container-title} a partir de um DOI.
}
