% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descobre_doi_por_titulo.R
\name{descobre_doi_por_titulo}
\alias{descobre_doi_por_titulo}
\title{Encontra o DOI via titulo do artigo}
\usage{
descobre_doi_por_titulo(titulo)
}
\arguments{
\item{titulo}{titulo do artigo}
}
\value{
DOI do artigo
}
\description{
Encontra o DOI via titulo do artigo
}
